Move all of the initializer code out to an
Initializer module.
